bypass pydantic json serialize/deserialize completely for /groups/public
the /groups/public api is called so often, and with very few differences
in parameters, that we want to cache the calculated results for 60-90
seconds (based on spoken_langauge), but the public group list can be
very large, e.g. the english+german list in production is around 260kb
last I checked, which takes hundreds of ms to serialize to pydantic
objects then to deserialize again to send the http response. Bypass this
whole thing to make this api a lot quicker. This also includes the
addition of the orjson dependency that is a much faster json library
than the built in json one.